Allow multiple TS players to control the same native player. (#456)
## Problem (PRN-122)
Some clients implement fullscreen by creating a new fullscreen Player instance
with the name nativeId.
This was documented as supported, but on Android, was not.

## Changes
When a native ID is reused, don't do initialize a new native instance.
Instead, reuse the same native instance.

## Test
Have 2 players on the same screen with the same native-id. Pausing one pauses the other.