Add DecoderConfig to Player (#622)
## Description
With the player-android `3.106` release we will introduce a new
`DecoderConfig` to the `PlaybackConfig`.
With this PR we introduce the `DecoderConfig` API on react-native.

In addition, the `3.106` release will remove some Tweaks. The affected
tweak will also be removed in this react-native project in favor of the
`DecoderConfig` (Removed tweak: `preferSoftwareDecodingForAds`)

## Changes
- Android: Add `DecoderConfigModule`
- RN: Add `decoderConfig` (+ usage in `PlaybackConfig`)
- RN: Remove `TweaksConfig.preferSoftwareDecodingForAds`

## Checklist
- [x] 🗒 `CHANGELOG` entry

---------

Co-authored-by: Roland Kakonyi <roland.kakonyi@bitmovin.com>
Co-authored-by: Žiga Vehovec <ziga.vehovec@bitmovin.com>